+++ idf_build_apps/app.py
@@ -19,7 +19,6 @@ from . import (
     LOGGER,
 )
 from .constants import (
-    DEFAULT_SDKCONFIG,
     IDF_PY,
     IDF_SIZE_PY,
     IDF_VERSION,
@@ -33,16 +32,14 @@ from .utils import (
     BuildError,
     dict_from_sdkconfig,
     find_first_match,
+    get_sdkconfig_defaults,
     rmdir,
     subprocess_run,
     to_list,
 )
 
 try:
-    from typing import (
-        Pattern,
-        TextIO,
-    )
+    import typing as t
 except ImportError:
     pass
 
@@ -56,10 +53,6 @@ class App:
 
     BUILD_SYSTEM = 'unknown'
 
-    SDKCONFIG_LINE_REGEX = re.compile(r"^([^=]+)=\"?([^\"\n]*)\"?\n*$")
-
-    SIZE_JSON = 'size.json'
-
     # could be assigned later, used for filtering out apps by supported_targets
     MANIFEST = None  # type: Manifest | None
 
@@ -68,21 +61,22 @@ class App:
     # Log this many trailing lines from a failed build log, also
     LOG_DEBUG_LINES = 25
     # IGNORE_WARNING_REGEX is a regex for warnings to be ignored. Could be assigned later
-    IGNORE_WARNS_REGEXES = []  # type: list[Pattern]
+    IGNORE_WARNS_REGEXES = []  # type: list[t.Pattern]
 
     def __init__(
         self,
-        app_dir,
-        target,
-        sdkconfig_path=None,
-        config_name=None,
-        work_dir=None,
-        build_dir='build',
-        build_log_path=None,
-        size_json_path=None,
-        check_warnings=False,
-        preserve=True,
-    ):  # type: (str, str, str | None, str | None, str | None, str, str | None, str | None, bool, bool) -> None
+        app_dir,  # type: str
+        target,  # type: str
+        sdkconfig_path=None,  # type: str | None
+        config_name=None,  # type: str | None
+        work_dir=None,  # type: str | None
+        build_dir='build',  # type: str
+        build_log_path=None,  # type: str | None
+        size_json_path=None,  # type: str | None
+        check_warnings=False,  # type: bool
+        preserve=True,  # type: bool
+        sdkconfig_defaults_list=None,  # type: list[str] | None
+    ):  # type: (...) -> None
         # These internal variables store the paths with environment variables and placeholders;
         # Public properties with similar names use the _expand method to get the actual paths.
         self._app_dir = app_dir
@@ -93,6 +87,11 @@ class App:
 
         self.name = os.path.basename(os.path.realpath(app_dir))
         self.sdkconfig_path = sdkconfig_path
+        self.sdkconfig_defaults_list = (
+            sdkconfig_defaults_list
+            if sdkconfig_defaults_list is not None
+            else get_sdkconfig_defaults(app_dir)  # use default
+        )
         self.config_name = config_name
         self.target = target
         self.check_warnings = check_warnings
@@ -127,6 +126,7 @@ class App:
             'build_dir',
             'size_json_path',
             'sdkconfig_path',
+            'sdkconfig_defaults_list',
             'config_name',
             'target',
             'check_warnings',
@@ -147,6 +147,7 @@ class App:
                 self._size_json_path,
                 self.name,
                 self.sdkconfig_path,
+                frozenset(self.sdkconfig_defaults_list),
                 self.config_name,
                 self.target,
                 self.check_warnings,
@@ -223,99 +224,37 @@ class App:
 
         return None
 
-    def build_prepare(self):  # type: () -> dict[str, str]
-        LOGGER.debug('=> Preparing Folders')
-        if self.work_dir != self.app_dir:
-            if os.path.exists(self.work_dir):
-                LOGGER.debug('==> Work directory %s exists, removing', self.work_dir)
-                if not self.dry_run:
-                    shutil.rmtree(self.work_dir)
-            LOGGER.debug('==> Copying app from %s to %s', self.app_dir, self.work_dir)
-            if not self.dry_run:
-                shutil.copytree(self.app_dir, self.work_dir)
-
-        if os.path.exists(self.build_path):
-            LOGGER.debug('==> Build directory %s exists, removing', self.build_path)
-            if not self.dry_run:
-                shutil.rmtree(self.build_path)
-
-        if not self.dry_run:
-            os.makedirs(self.build_path)
-
-        # Prepare the sdkconfig file, from the contents of sdkconfig.defaults (if exists) and the contents of
-        # build_info.sdkconfig_path, i.e. the config-specific sdkconfig file.
-        #
-        # Note: the build system supports taking multiple sdkconfig.defaults files via SDKCONFIG_DEFAULTS
-        # CMake variable. However here we do this manually to perform environment variable expansion in the
-        # sdkconfig files.
-        sdkconfig_defaults_list = [
-            'sdkconfig.defaults',
-            'sdkconfig.defaults.' + self.target,
-        ]
-        if self.sdkconfig_path:
-            sdkconfig_defaults_list.append(self.sdkconfig_path)
-
-        LOGGER.debug('=> Generating sdkconfig file')
-        sdkconfig_file = os.path.join(self.work_dir, 'sdkconfig')
-        if os.path.exists(sdkconfig_file):
-            LOGGER.debug('==> Removing sdkconfig file: %s', sdkconfig_file)
-            if not self.dry_run:
-                os.unlink(sdkconfig_file)
-
-        LOGGER.debug('==> Creating sdkconfig file: %s', sdkconfig_file)
-        cmake_vars = {}
-        if not self.dry_run:
-            with open(sdkconfig_file, 'w') as f_out:
-                for sdkconfig_name in sdkconfig_defaults_list:
-                    sdkconfig_path = os.path.join(self.work_dir, sdkconfig_name)
-                    if not sdkconfig_path or not os.path.exists(sdkconfig_path):
-                        continue
-                    LOGGER.debug('==> Appending %s to sdkconfig', sdkconfig_name)
-                    with open(sdkconfig_path, 'r') as f_in:
-                        for line in f_in:
-                            if not line.endswith('\n'):
-                                line += '\n'
-                            if isinstance(self, CMakeApp):
-                                m = self.SDKCONFIG_LINE_REGEX.match(line)
-                                key = m.group(1) if m else None
-                                if key in self.SDKCONFIG_TEST_OPTS:
-                                    cmake_vars[key] = m.group(2)
-                                    continue
-                                if key in self.SDKCONFIG_IGNORE_OPTS:
-                                    continue
-                            f_out.write(os.path.expandvars(line))
-        else:
-            for sdkconfig_name in sdkconfig_defaults_list:
-                sdkconfig_path = os.path.join(self.work_dir, sdkconfig_name)
-                if not sdkconfig_path:
-                    continue
-                LOGGER.debug('==> Considering sdkconfig %s', sdkconfig_path)
-                if not os.path.exists(sdkconfig_path):
-                    continue
-                LOGGER.debug('==> Appending %s to sdkconfig', sdkconfig_name)
-
-        return cmake_vars
-
     @classmethod
-    def _get_default_sdkconfig_target(cls, path):  # type: (str) -> str | None
-        # check if there's CONFIG_IDF_TARGET in sdkconfig.defaults
+    def _get_sdkconfig_defaults_defined_idf_target(
+        cls,
+        path,  # type: str
+        sdkconfig_defaults_list,  # type: list[str]
+    ):  # type: (...) -> str | None
+        """
+        Check if there's CONFIG_IDF_TAREGET defined in sdkconfig defaults files
+        """
         default_sdkconfig_target = None  # type: str | None
-        default_sdkconfig = os.path.join(path, DEFAULT_SDKCONFIG)
-        if os.path.isfile(default_sdkconfig):
-            sdkconfig_dict = dict_from_sdkconfig(default_sdkconfig)
-            if 'CONFIG_IDF_TARGET' in sdkconfig_dict:
-                default_sdkconfig_target = sdkconfig_dict['CONFIG_IDF_TARGET']
-                LOGGER.debug(
-                    'CONFIG_IDF_TARGET is set in %s. Set enable build targets to %s only.',
-                    default_sdkconfig,
-                    default_sdkconfig_target,
-                )
+        for f in sdkconfig_defaults_list:
+            fp = os.path.join(path, f)
+            if os.path.isfile(fp):
+                sdkconfig_dict = dict_from_sdkconfig(fp)
+                if 'CONFIG_IDF_TARGET' in sdkconfig_dict:
+                    default_sdkconfig_target = sdkconfig_dict['CONFIG_IDF_TARGET']
+                    LOGGER.debug(
+                        'CONFIG_IDF_TARGET is set in %s. Set enable build targets to %s only.',
+                        fp,
+                        default_sdkconfig_target,
+                    )
 
         return default_sdkconfig_target
 
     @classmethod
-    def enable_build_targets(cls, path):  # type: (str) -> list[str]
-        default_sdkconfig_target = cls._get_default_sdkconfig_target(path)
+    def enable_build_targets(
+        cls,
+        path,  # type: str
+        sdkconfig_defaults_list,  # type: list[str]
+    ):  # type: (...) -> list[str]
+        default_sdkconfig_target = cls._get_sdkconfig_defaults_defined_idf_target(path, sdkconfig_defaults_list)
         if cls.MANIFEST:
             res = cls.MANIFEST.enable_build_targets(path, default_sdkconfig_target)
         else:
@@ -327,8 +266,12 @@ class App:
         return res
 
     @classmethod
-    def enable_test_targets(cls, path):  # type: (str) -> list[str]
-        default_sdkconfig_target = cls._get_default_sdkconfig_target(path)
+    def enable_test_targets(
+        cls,
+        path,  # type: str
+        sdkconfig_defaults_list,  # type: list[str]
+    ):  # type: (...) -> list[str]
+        default_sdkconfig_target = cls._get_sdkconfig_defaults_defined_idf_target(path, sdkconfig_defaults_list)
 
         if cls.MANIFEST:
             return cls.MANIFEST.enable_test_targets(path, default_sdkconfig_target)
@@ -344,11 +287,11 @@ class App:
 
     @property
     def supported_targets(self):
-        return self.enable_build_targets(self.app_dir)
+        return self.enable_build_targets(self.app_dir, self.sdkconfig_defaults_list)
 
     @property
     def verified_targets(self):
-        return self.enable_test_targets(self.app_dir)
+        return self.enable_test_targets(self.app_dir, self.sdkconfig_defaults_list)
 
     @abstractmethod
     def build(
@@ -384,7 +327,7 @@ class App:
         )
         LOGGER.info('=> Generated size info to %s', self.size_json_path)
 
-    def collect_size_info(self, output_fs):  # type: (TextIO) -> None
+    def collect_size_info(self, output_fs):  # type: (t.TextIO) -> None
         if not os.path.isfile(self.size_json_path):
             self.write_size_json()
 
@@ -426,19 +369,13 @@ class App:
 
         return True, is_ignored
 
+    @classmethod
+    def is_app(cls, path):  # type: (str) -> bool
+        raise NotImplementedError('Please implement this function in sub classes')
+
 
 class CMakeApp(App):
     BUILD_SYSTEM = 'cmake'
-
-    # If these keys are present in sdkconfig.defaults, they will be extracted and passed to CMake
-    SDKCONFIG_TEST_OPTS = [
-        'EXCLUDE_COMPONENTS',
-        'TEST_EXCLUDE_COMPONENTS',
-        'TEST_COMPONENTS',
-    ]
-
-    # These keys in sdkconfig.defaults are not propagated to the final sdkconfig file:
-    SDKCONFIG_IGNORE_OPTS = ['TEST_GROUPS']
 
     # While ESP-IDF component CMakeLists files can be identified by the presence of 'idf_component_register' string,
     # there is no equivalent for the project CMakeLists files. This seems to be the best option...
@@ -449,29 +386,70 @@ class CMakeApp(App):
         depends_on_components=None,  # type: list[str] | str | None
         check_component_dependencies=False,  # type: bool
     ):  # type: (...) -> bool
-        cmake_vars = self.build_prepare()
+        LOGGER.debug('=> Preparing...')
+        if self.work_dir != self.app_dir:
+            if os.path.exists(self.work_dir):
+                LOGGER.debug('==> Work directory %s exists, removing', self.work_dir)
+                if not self.dry_run:
+                    shutil.rmtree(self.work_dir)
+            LOGGER.debug('==> Copying app from %s to %s', self.app_dir, self.work_dir)
+            if not self.dry_run:
+                shutil.copytree(self.app_dir, self.work_dir)
+
+        if os.path.exists(self.build_path):
+            LOGGER.debug('==> Build directory %s exists, removing', self.build_path)
+            if not self.dry_run:
+                shutil.rmtree(self.build_path)
+
+        if not self.dry_run:
+            os.makedirs(self.build_path)
+
+        sdkconfig_file = os.path.join(self.work_dir, 'sdkconfig')
+        if os.path.exists(sdkconfig_file):
+            LOGGER.debug('==> Removing sdkconfig file: %s', sdkconfig_file)
+            if not self.dry_run:
+                os.unlink(sdkconfig_file)
 
         if self.build_log_path:
             LOGGER.info('=> Writing build log to %s', self.build_log_path)
+
+        if self.dry_run:
+            LOGGER.debug('==> Skipping... (dry run)')
+            return True
+
+        if self.build_log_path:
             log_file = open(self.build_log_path, 'w')
         else:
             # delete manually later, used for tracking debugging info
             log_file = tempfile.NamedTemporaryFile('w', delete=False)
 
+        # additional env variables
+        # IDF_TARGET to bypass the idf.py build check
+        # SDKCONFIG_DEFAULTS to bypass the file not exists issue
+        additional_env_dict = {
+            'IDF_TARGET': self.target,
+            'SDKCONFIG_DEFAULTS': ';'.join(self.sdkconfig_defaults_list) if self.sdkconfig_defaults_list else ';',
+        }
+
+        # check if this app depends on components according to the project_description.json 'build_component' field.
+        # the file is generated by `idf.py reconfigure`.
+        common_args = [
+            sys.executable,
+            str(IDF_PY),
+            '-B',
+            self.build_path,
+            '-C',
+            self.work_dir,
+        ]
+
         depends_on_components = to_list(depends_on_components)
         if depends_on_components is not None and check_component_dependencies:
-            reconfigure_args = [
-                sys.executable,
-                str(IDF_PY),
-                '-B',
-                self.build_path,
-                '-C',
-                self.work_dir,
-                '-DIDF_TARGET=' + self.target,
-                'reconfigure',
-            ]
             subprocess_run(
-                reconfigure_args, log_terminal=False if self.build_log_path else True, log_fs=log_file, check=True
+                common_args + ['reconfigure'],
+                log_terminal=False if self.build_log_path else True,
+                log_fs=log_file,
+                check=True,
+                additional_env_dict=additional_env_dict,
             )
 
             with open(os.path.join(self.build_path, PROJECT_DESCRIPTION_JSON)) as fr:
@@ -486,44 +464,18 @@ class CMakeApp(App):
                 )
                 return False
 
-        build_args = [
-            sys.executable,
-            str(IDF_PY),
-            '-B',
-            self.build_path,
-            '-C',
-            self.work_dir,
-            '-DIDF_TARGET=' + self.target,
-        ]
-        if cmake_vars:
-            for key, val in cmake_vars.items():
-                build_args.append('-D{}={}'.format(key, val))
-            if 'TEST_EXCLUDE_COMPONENTS' in cmake_vars and 'TEST_COMPONENTS' not in cmake_vars:
-                build_args.append('-DTESTS_ALL=1')
-            if 'CONFIG_APP_BUILD_BOOTLOADER' in cmake_vars:
-                # In case if secure_boot is enabled then for bootloader build need to add `bootloader` cmd
-                build_args.append('bootloader')
-
-        build_args.append('build')
-
+        # idf.py build
+        build_args = common_args + ['build']
         if self.verbose:
             build_args.append('-v')
-
-        if self.dry_run:
-            LOGGER.debug('==> Skipping... (dry run)')
-            return True
-
-        old_idf_target_env = os.getenv('IDF_TARGET')
-        os.environ['IDF_TARGET'] = self.target  # pass the cmake check
 
         returncode = subprocess_run(
             build_args,
             log_terminal=False if self.build_log_path else True,
             log_fs=log_file,
+            additional_env_dict=additional_env_dict,
         )
 
-        if old_idf_target_env is not None:
-            os.environ['IDF_TARGET'] = old_idf_target_env  # revert it back
         log_file.close()
 
         # help debug