+++ idf_build_apps/main.py
@@ -44,7 +44,7 @@ def find_apps(
     exclude_list=None,  # type: list[str] | None
     work_dir=None,  # type: str | None
     build_dir='build',  # type: str
-    config_rules_str=None,  # type: list[str] | None
+    config_rules_str=None,  # type: list[str] | str | None
     build_log_path=None,  # type: str | None
     size_json_path=None,  # type: str | None
     check_warnings=False,  # type: bool
@@ -55,6 +55,7 @@ def find_apps(
     manifest_rootpath=None,  # type: str | None
     ignore_component_dependencies_file_patterns=None,  # type: list[str] | str | None
     depends_on_files=None,  # type: list[str] | str | None
+    sdkconfig_defaults=None,  # type: str | None
 ):  # type: (...) -> list[App]
     """
     Find app directories in paths (possibly recursively), which contain apps for the given build system, compatible
@@ -75,7 +76,7 @@ def find_apps(
     :param build_dir: directory where the build will be done. Support placeholders.
     :type build_dir: str
     :param config_rules_str: mapping of sdkconfig file name patterns to configuration names
-    :type config_rules_str: list[str] | None
+    :type config_rules_str: list[str] | str | None
     :param build_log_path: path of the build log. Support placeholders.
         The logs will go to stdout/stderr if not specified
     :type build_log_path: str | None
@@ -102,6 +103,9 @@ def find_apps(
     :param depends_on_files: skip check app's component dependencies if any of the specified files matches
         ``ignore_component_dependencies_file_patterns``
     :type depends_on_files: list[str] | str | None
+    :param sdkconfig_defaults: semicolon-separated string, pass to idf.py -DSDKCONFIG_DEFAULTS if specified,
+        also could be set via environment variables "SDKCONFIG_DEFAULTS"
+    :type sdkconfig_defaults: str | None
     :return: list of found apps
     :rtype: list[App]
     """
@@ -162,6 +166,7 @@ def find_apps(
                     preserve=preserve,
                     depends_on_components=depends_on_components,
                     check_component_dependencies=check_component_dependencies,
+                    sdkconfig_defaults_str=sdkconfig_defaults,
                 )
             )
     apps.sort()