+++ CHANGELOG.md
@@ -4,11 +4,21 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Support specify `-DSDKCONFIG_DEFAULTS` for `idf.py build`
+  - via CLI option `--sdkconfig-defaults`
+  - via environment variable `SDKCONFIG_DEFAULTS`
+
 ### Fixed
 
 - Stop writing `app_info` and `size_info` if the build got skipped
 - `IDF_VERSION_MAJOR`, `IDF_VERSION_MINOR`, `IDF_VERSION_PATCH` now are integers
 - Skip exclude files while removing build directory if files not exist
+
+### Breaking Changes
+
+- Environment variable expansions in sdkconfig files is no longer supported
 
 ## [0.4.0]